[IDA-812] Add support for post bodies
AciParameter converts its values to Strings, which is inappropriate for
InputStream parameters. We define a new parameter interface,
ActionParameter<T>, which represents a parameter. AciParameter has been
retrofitted to this interface, and InputStreamActionParameter, which
allows InputStreams to be added as parameters. In the presence of such
a parameter, the AciHttpClient will post each parameter to IDOL as a
single part of a multipart request. Otherwise, the existing behaviour
is maintained. The ActionParameters helper class has been added to
handle InputStreamActionParameter construction.

Deprecate AciParameters as the only reason to prefer it is the fact that
it implements Serializable.

Note that this changes the AciService and AciHttpClient interfaces in a
backwards incompatible manner, due to the new methods having the same
erasure as the old ones.

[rev: matthew.gordon]